backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.7.0
pymongo>=4.7.0
google-generativeai>=0.5.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
email-validator>=2.1.1
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
